backend/requirements.txt
# Base installation
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

# System
python-dotenv==1.0.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Data
pandas==2.2.2
scikit-learn==1.5.1
Unidecode==1.3.8
textblob==0.18.0.post0
vaderSentiment==3.3.2

# Api
newsapi-python==0.2.7
requests==2.32.3
drf-spectacular==0.27.2

# NLP
nltk==3.8.1
spacy==3.7.5
gensim==4.3.2

# DevOps
gunicorn==22.0.0